Add parallel processing with dask.distributed (#16)
Add command line arguments `--dask-distributed-local-core-fraction` and `--dask-distributed-local-memory-fraction` which can be used to enable multiprocessing by setting the former to a non-zero fraction representing the fraction of CPU cores on the local machine to use for parallel processing with `dask.distrubted.LocalCluster`. The latter argument set what fraction of the total system memory is allocated to the workers.