add section for extra info in config (#18)
Add optional section to config called `extra` that allows the user to include arbitrary information that is ignored by mllam-data-prep. This is added so that this extra information can be used in downstream applications (for example neural-lam will use this for projection info for now, until we are able to complete setting projection info correctly in a CF-compliant manner).