Reworked how to define statistical methods
Each statistical method now have its own class with a calc_stats method. In the config yaml file, one specifies the name of the classes to use when computing statistics of the dataset. If the class is not defined in the statistics.py

Added classes for calculating Diurnal diff mean, and Diurnal diff std, and Diff time mean (where only the time is averaged - not the grid_index)