Use CF-compliant coordinate stacking rather than .reset_index()
This commit replaces the current stacking method where
`xr.Dataset.reset_index()` is used to change the stacked coordinate (for
example `grid_index` as a stacking of `x` and `y` coordinates) from a
`pd.MultiIndex` to a regular index, and instead uses the CF-compliant "gather
compression" of coordinate stacking (see
https://cf-xarray.readthedocs.io/en/latest/coding.html). Using `reset_index()`
is a lossy operation in that downstream applications will not know wwhich
coordinates have been stacked, meaning we must instead rely on this information
from elsewhere (e.g. the config) in downstream applications.
The reason that stacked coordinates must be handled in a special way is that
`pd.MultiIndex` coordinates cannot be directly written to netCDF/zarr datasets,
but with the CF-compliant method they can be handled in such a manner that we
can completely roundtrip in downstream applications.